Use always RFN when handling RACH indications
The previous code was really confusing, passing full FNs as RFNs under
certain external conditions, and then assuming the RFN input of
rfn_to_fn() function could actually be a FN.

As a result, we had a lot of code behaving slightly different depending
on whether the incomding FN from pcuif was filled in by a BSC or a BTS.
Avoid this b ehavior differentiation and always assume the most
restricted one, aka RFN.

+++ src/bts.h
@@ -196,7 +196,7 @@ struct rach_ind_params {
 	uint16_t ra;
 	uint8_t trx_nr;
 	uint8_t ts_nr;
-	uint32_t rfn;
+	uint16_t rfn;
 	int16_t qta;
 };
